Fixed issues with multi-Lisp book compilation from books/GNUmakefile.  Fixed some issues with must-fail.  Synched doc.lisp.
Made some updates for the compilation of books in more than one Lisp
(books/GNUmakefile invoked with ACL2_COMP=t, which is an obscure
mechanism that is probably used only by me).

Updated books/misc/eval.lisp to restore some behavior from Version
7.1.  Basically, must-fail and must-succeed do a more appropriate
passing along of the :check-expansion argument.  It is now possible to
evaluate the form
(include-book "make-event/eval-check-tests" :dir :system)
successfully, as was the case in Version 7.1 but had since no longer
been the case (till now).

Updated books/kestrel/general/testing-tests.lisp to avoid failure to
be able to include the book after certification.
NOTE: The rest of this check-in message describes the situation there.

The problem was that must-fail doesn't force the proofs to be
attempted when proofs are skipped, in particular when the book is
included, so a form (defthm ...) that is expected to fail will
unfortunately succeed during include-book.  My solution was to make
three forms (must-fail (defthm ...)) local so that they're skipped
during include-book.

NOTE: A different solution is to add the form

(include-book "make-event/eval-check" :dir :system)

either to testing-tests.lisp or to the book it already includes, and
then change each of the three events of the form

(must-fail (defthm ...))

to

(must-fail! (defthm ...))

since must-fail! causes proofs to be attempted even during
include-book.