quadratic-reciprocity/euclid: Add guards
Add guards to projects/quadratic-reciprocity/euclid.lisp because this book is used
in another project which respects guards.

Guards for predicates are true.
This caused to add conjuncts in defintion of predicate divides
and to use (equal ...) instead of (= ...) in definition of predicate primep.

Weaker hypothesis are possible in theorems about divides now.