Meaner stobj-swap tests that provoke CMUCL problem
Extend stobj-swap-tests with some additional tests that expose a bad
interaction between def-stobj-swap and (comp t).  I'm not sure how we
should fix this yet, but at least this clearly captures what the
problem is.

I think this captures the root of the problem that causes
centaur/sv/tutorial/boothpipe.lisp to fail to certify on CMUCL and
perhaps other Lisps.

In particular, in that file, SWAP-ELAB-MODS seems to have the wrong
definition before the (defsvtv boothpipe-direct).  It seems that
what's happening here is:

   - We load centaur/sv/mods/moddb.lisp via "../top"
   - We then load centaur/misc/defapply.lisp via "centaur/gl/gl"
   - This defapply book submits a (comp t), ruining SWAP-ELAB-MODS.