x86isa: Generalized rules about wb by changing its first argument in those rules from (create-addr-bytes-alist ... ...) to a variable.
Also proved wb-wb-split-reads-in-system-level-mode.

Note that all the rules that rewrite rm*/wm* to rb/wb, except those
for rm08 and wm08, are currently broken. I think I'll prove these
lemmas by using rb-rb-split-reads-in-system-level-mode and
wb-wb-split-reads-in-system-level-mode, after I figure out whether the
definitions of rm*/wm* functions in the system-level mode are indeed
equal to the repeated operations of rm08/wm08, respectively.

To see why this is an issue, consider wm16. If wm16 were indeed equal
two successive wm08s, then there's the problem of atomicity. What if
there's an error in the translation of the linear address of the
second write and not the first? Do we undo the effects of the first
wm08, but keep the dirty/accessed bits intact, or do we also undo the
updates to dirty/accessed bits? Or do we update accessed/dirty bits
according as though writes of 8 bytes have taken place, because maybe
x86 does that? Sigh.