SV: Big change to how the rewriter works.
Remove the localp flag from rewrite rules and remove the local rewriting
pass on the arguments of a function call.  This wasn't giving much
benefit and was taking a lot of time (maybe 30% of the total rewriting
time).

Instead, track which subexpressions are referenced more than once.  A
reasonable guess is that this is around 10% or less of the total number
of subexpressions.  Non-multiply-referenced subexpressions can be
rewritten less carefully than multiply-referenced ones.  We don't yet
have rules that take advantage of this much, but we have that
opportunity now.  Additionally, we manually memoize svex-rewrite and
only record the results for multiply-referenced nodes.