x86isa: Resolved the question of atomicity of memory reads and writes in the system-level mode
In the last few commits, there was some question of whether the
definitions of rm*/wm* functions in the system-level mode are indeed
equal to the repeated operations of rm08/wm08 or not.

We've resolved this issue by making the following changes: an
operation like wm16 is not equal to two successive wm08s. Instead, two
address translation operations occur first; if these two operations
are error-free, then an "atomic" 2-byte write to the memory is
done. No partial writes to the memory occur.

Note that we already had this definition for all rm*/wm* functions
except 8-byte and 16-byte operations, which have now been modified
accordingly.