VL: some enum fixes and new enum tests
Fix two bugs with weird enum ranges, i.e., enums with items like bar
and baz in:

    enum logic [3:0] {
       foo,
       bar[3],
       baz[3:0],
    };

Bug 1: the parser was treating bar[3] as bar[0:3], but it should
actually be bar[0:2].

Bug 2: the enumnames transform was screwing up the assignments to
enums with ranges by forgetting to tack on the name index in the
nextval expression.

Add two new cosims that do some basic testing of enum ranges.

Add a failtest to make sure enum items like bar[0] are rejected.