x86isa: Rewrite all rm*/wm* calls to rb/wb, when possible, in the system-level mode
The x86isa books provide various top-level linear memory read (rm08
through rm128) and write (wm08 through wm128,
write-canonical-address-to-memory) functions. During reasoning, these
functions simply open up to rb and wb in the programmer-level mode,
thus giving us a canonical way of looking at memory accesses and
updates. However, in the system-level mode, it's not so simple because
well-formedness of paging structures should also be considered.

The motivation for this change is to make reasoning about memory in
system-level mode close to what it is in the programmer-level
mode. The new rules that rewrite rm*/wm* to rb/wb in the system-level
mode have the well-formedness of the paging structures as the
pre-condition. We force the hypotheses of these rules because we will
almost always be reasoning in a state where the paging structures are
well-formed.