x86isa: Finished the proof of wb-remove-duplicate-writes-in-system-level-mode.
This new rule removes shadowed writes from a call of wb in the
system-level mode. For example:

 (wb '((0 . 1) (1 . 2) (0 . 5)) x86)

would be rewritten to

 (wb '((1 . 2) (0 . 5)) x86).

The first argument of wb is an alist that maps linear addresses to
bytes; this alist has the most recent update as its last element. The
new lemma can actually do better than simply removing stale writes to
a linear address; it can remove stale writes to a physical address
too. Suppose in our example the linear addresses 0 and 1 were mapped
to the same physical address. Then, the original call will be
re-written to

(wb '((0 . 5)) x86).