Replace EXECUTABLEP system utility with NON-EXECUTABLEP.
As pointed out by Matt Kaufmann, a NON-EXECUTABLEP property with value NIL is
not the same as the function being "executable" in an intuitive sense. It just
means that the function has not been explicitly marked as non-executable
(e.g. via a DEFUN-NX). So the new system utility just returns the value of that
property; its documentation clarifies the relationship between the result and
"executability".