VL: sizing bug fix, lint tweak
There was a bug with our handling of extension integers like '1, 'x,
etc.  In particular, we were treating them as signed, but they are
supposed to be unsigned.  Fix the bug and add some more cosims to test
it more.

Tweak basicsanity check to restore an (elided) description of the
offending port in port-style warnings, which is good enough to get
the linttests passing again.