x86isa: Pushed through the proof of correctness of rewireVirtualAddress (zero copy).
Right now, I've just proved that the first quad-word of the
destination in the final x86 state is equal to the first quad-word of
the source in the initial x86 state.  Note that this proof has been
done in the system-level non-marking mode, i.e., supervisor mode where
paging structure walks do not set the accessed and/or dirty bits.

Things to do as far as this proof is concerned:

1. Clean up this proof!! Seriously, it's such a big mess right now.
2. Extend this proof to 1G of data, instead of just a quad-word.

Time permitting, I might attempt to prove this program correct in the
system-level marking mode.