Strengthened new type-set heuristics, in part to strengthen bitp-compound-recognizer for the case that (bitp term) is false.
Thanks to Jared Davis for pointing out that our first attempt to add a
new type-set bit for the singleton set {1} left
bitp-compound-recognizer weaker than expected.

The new mods slowed things down slightly on my Mac, but I'm hopeful
that the price is only an additional percent or so on Linux, where
timings are more reliable; feel free to query the list tomorrow
(Friday) about that.  A tiny bit of slowdown seems appropriate to me;
I think that often happens with additional reasoning power, where more
is done, so we get a bit more coverage but some already-successful
proofs may slow down a bit.  I see this in particular when
free-variable matching for rewrite rule hypotheses is involved.