x86isa: Proved that two page walks are independent, given only that the x86 state is well-formed.
This proof was made much, much easier by over-approximating the list
of addresses that correspond to the paging data structures, as opposed
to collecting the exact set of addresses (see function
gather-all-paging-structure-qword-addresses). Collecting the exact set
of addresses required checking whether a superior paging structure
entry has PS = 0 and P = 1, whereas the over-approximation doesn't
make those checks.

This relaxation helped remove PS = 0 and P = 1 hypotheses cluttering
this independence proof. The equivalence relation
all-mem-except-paging-structures-equal always holds for the x86 state
obtained after a page walk because the only addresses modified during
the page walk belong to this over-approximated set (which, by
definition, is disjoint from the set considered in
all-mem-except-paging-structures-equal). The equivalence relation
xlate-equiv-structures also always holds for the x86 state obtained
after a page walk because the page walk always modifies entries such
that they satisfy xlate-equiv-entries (which includes equality as
well). Together, the state obtained after a page walk satisfies
xlate-equiv-memory (xlate-equiv-memory-with-mv-nth-2-ia32e-la-to-pa),
which, along with xlate-equiv-memory-and-mv-nth-0-ia32e-la-to-pa and
xlate-equiv-memory-and-mv-nth-1-ia32e-la-to-pa, helped prove
two-page-walks-ia32e-la-to-pa, given just that the x86 state is
well-formed.