All-ffn-symbs and all-ffn-symbs-lst are in :logic mode.  Use ens-from-pspv more in sources.
Quoting a comment in (defxdoc note-7-3 ...):

; Functions all-ffn-symbs and all-ffn-symbs-lst are now in :logic mode.  Thanks
; to Eric Smith for a helpful relevant conversation.

Function ens-from-pspv is a nice abbreviation that deserved more usage
in the sources, so I've remedied that (but I avoided one such change
on purpose where it would have perhaps decreased clarity).

By the way, I added to the end of the page

http://www.cs.utexas.edu/users/moore/acl2/open-architecture/how-to-make-patches.txt

to clarify how to put system functions into :logic mode.