Improved :doc system-utilities explanation of recursivep.  Brought guard-verified status for functions from books/system/kestrel.lisp into the ACL2 build.
The explanation of recursivep in :doc system-utilities failed to point
out the role of definition rules with non-nil :install-body.  That has
been fixed.  Thanks to Alessandro Coglio and Eric Smith for a helpful
chat leading to this improvement.

Functions that are guard-verified in books/system/kestrel.lisp now
come up as guard-verified when ACL2 starts up.