Remove ENSURE and ENSURE$.
These do not work well for logic-mode, guard-verified code, because they do not
enrich the context with the truth of the conditions they check. Instead, WHEN/IF
and UNLESS can be used in B*, with just slightly more verbosity. B* binders
corresponding to ENSURE and ENSURE$ could be also defined, if desired.