Add new system utility.
FRESH-NAME-IN-WORLD-WITH-$S adds zero or more $s to the given name to make it
new in the world, and also different from all the names in a given list.

This utility is under a new XDOC topic FRESH-NAMES, to collect this and other
similar utilities to generate fresh names (of functions, variables, etc.).