Add general utility (written by Jared Davis).
DEFINE-SK is a DEFINE-like version of DEFUN-SK. Besides the usual DEFINE-like
features (e.g. extended formals and XDOC integration), it facilitates guard
verification by regarding certain uses of IMPLIES as a non-strict construct
(like IF) for the purpose of verifying guards. For FORALL functions, it also
automatically generates the :DIRECT rewrite rule (when that formula is a valid
rewrite rule), which is generally more useful than the :DEFAULT rewrite rule.