Extend two term utilities.
The CHECK-USER-TERM and CHECK-USER-LAMBDA-EXPR term utilities now return, if
successful, not only the translated term or lambda expression, but also the
output stobjs of the term or (body of the) lambda expression.

Tests have been added for this new feature.

Additional tests for malformed lambda expressions have also been added.