Improve applicability condition utilities.
Now PROVE-APPLICABILITY-CONDITION and PROVE-APPLICABILITY-CONDITIONS return a
more clear result, consisting of a boolean to indicate success or failure, a
message, and the state. The message is an error message in case of failure, and
currently empty in case of success; but in the future it could carry information
also in case of success.