Added eager sharing for iprinting.  Fixed ACL2(p) bug due to earlier load of hons-raw.lisp during build.  Removed books/misc/check-acl2-exports.cert from ordinary regressions.
Quoting :doc note-7-3 -- this is an existing item that I modified to
mention eager sharing.

  The iprinting utility has a new keyword option, :share, which causes
  iprint indices to be re-used (using so-called ``iprint sharing''
  and ``iprint eager sharing''. See [set-iprint]. Thanks to David
  Rager for suggesting a notion of iprint sharing.

I also updated :doc set-iprint, not only to document eager sharing but
also to clarify how normal sharing works.

Thanks to David Rager for pointing out the compilation warning for
ACL2(p) builds that were occurring after loading hons-raw.lisp before
basis-a.lisp.

Note that books/misc/check-acl2-exports.cert will still be built in
"everything" regressions.  I removed this from ordinary regressions
because it was too easy for someone to break the build by documenting
a system function or macro symbol, since that requires either adding
that symbol to system constant *acl2-exports* or to the constant
*acl2-exports-exclusions* in books/misc/check-acl2-exports.lisp.  I
routinely do "everything" regressions, and I'm happy to have sole
responsibility for maintaining books/misc/check-acl2-exports.lisp,
which I think of as merely implementing a check on the system.

Clarified the binding of :stobjs-out in bindings returned by
translate1-cmp and translate11.  Thanks to Alessandro Coglio for a
query leading us to make this clarification.