x86isa: Fixed a bug when updating A/D flags during page walks if the paging entries are self-referential.
!! NOTE: Build is broken for this commit. paging/top is where the
   problem is.!!

Old behavior:

The A/D flags in a paging entry were set only when we detected that
they were zero.  Also, these flags were set from the innermost entry
to the outermost one.

For example, while mapping a linear address to a physical address
located in a 4K page, we would read the PML4TE e-1, the PDPTE e-2, the
PDE e-3, and the PTE e-4; then, we would set the A flags of e-4 (also
its D flag, if applicable), e-3, e-2, and then e-1, if they weren't
already set --- all in that order.

This worked fine as long as e-1, e-2, e-3, and e-4 were all distinct
entries.  However, self-referential paging entries are allowed by the
ISA, which means that just one entry could serve as all four e-1, e-2,
e-3, and e-4.  Unfortunately, the check whether the A flag wasn't
already set used the "stale" entry that did not reflect the A/D
updates of walking an inferior entry.  This could lead to incorrectly
clearing the D flag (updated by the PTE walk if the walk is on behalf
of a write) because the D flag is always clear/ignored in a paging
entry that does not map a page.  Note that working with a stale entry
only affected the D flag and not the A flag --- the walk of the
superior entry would update the A flag anyway.

New behavior:

Every function that walks a paging data structure obtains a fresh copy
of the paging entry under consideration before detecting/updating the
A/D flags, only if that entry refers to another paging structure
instead of mapping a page.  This is because if an entry maps a page,
it is the innermost (and thus, the freshest) entry already.  Note that
self-referential entries are possible only when the paging entries map
a 4K page --- a 4K page configuration is the only one which allows the
PS field of all the entries to be the same.