Made miscellaneous iprovements for assertion stuff.
Value-triple can specify a :ctx for an error and assert.

ERRORS is now a parent of ASSERT-EVENT (thanks to Eric Smith for the
suggestion).

VALUE-TRIPLE now takes a :ctx keyword, which is used in ASSERT-EVENT.

ASSERT! and ASSERT!-STOBJ fail without all the usual error noise from
make-event.

Fixed typo (missing colon) in :doc ld pointed out by Eric Smith.

Extended ASSERT!-STOBJ to work not only when the stobj is
user-defined, but also when it is STATE,