Fixed mechanisms for dealing with functions that are guard-verified in books.  Made remove-lambdas a guard-verified function.  Tweaked :doc verify.
Quoting a comment in :doc note-7-3:

; Fixed a bug in system-verify-guards-fn-1 that was failing to verify guards
; when the original defun specified xargs :verify-guards nil.  Also fixed a bug
; in check-none-ideal that was failing to consider certain event tuples, and
; thus was allowing :ideal mode functions (of the sort mentioned in the
; preceding sentence) to exist at start-up.

With this change, remove-lambdas is now a guard-verified function at
start-up.

Improved :doc verify to explain how EXIT can be used to save events,
and also to explain why that might not be a good idea.  Thanks to
Mihir Mehta for a query that led us to this enhancement.