Make fty::defprod with (unsigned-byte-p NNN x) fields faster.
fty::defprod allows type of field to be a macro.
However, defprod is slow in this case because
it can't find user-supplied fix-pred and pred-of-fix theorems.
Often this macro is expanded to (unsinged-byte-p NNN x)
where NNN is a natural constant.
THis commit modifies flextypes-collect-fix/pred-pairs-aux so that it can handle
this particular case.
General solution for any macro is difficult for me now.