Fixed bug in std::get-event-tuple.  Added :doc defund-nx.  Clarified time taken during installation for make certify-books.
Details:

  Fixed bug in std::get-event-tuple.

Thanks to Eric Smith for a query leading me to make this fix.

  Added :doc defund-nx.

Thanks to Eric Smith for suggesting this topic addition.

  Clarified time taken during installation for make certify-books.

Thanks to Charlotte Karako for pointing out that it can take longer
for make certify-books than suggested in the installation
instructions.  (In her case it was due to the use of CMU on (probably)
a significantly slower machine than I use.)