Improved the :termination-theorem lemma-instance when admitting definitions.  Added tool make-termination-theorem that creates a copy of a given termination theorem that is suitable for functional instantiation.
Quoting :doc note-7-3:

  The :[termination-theorem] [lemma-instance] has been improved so that
  old function symbols may be replaced with corresponding new
  function symbols, either automatically or explicitly by the user.
  See [lemma-instance] and see [termination-theorem-example]. Thanks
  for Alessandro Coglio and Eric Smith for helpful discussions,
  including pointing out the previous deficiency.

I plan soon to document make-termination-theorem (in the new community
book books/kestrel/utilities/make-termination-theorem.lisp).  I also
plan to add tests for the keyword arguments (which may expose bugs).
For now, an example at the end of the file gives a sense of how to use it.

Updates to :doc termination-theorem-example and :doc lemma-instance
discuss the new ability to apply a functional substitution to an
existing termination theorem.