Store proved functional instances from inside trivial encapsulates.  Fixed minor #f bug.  Added :quiet argument to write-list.
Quoting :doc note-7-3:

  For many years, it has been the case that when a [constraint] is
  proved that arises from a :functional-instance lemma instance (see
  [lemma-instance]), that information is stored so that the same
  constraint is not re-proved on behalf of a later such lemma
  instance. However, this had not been the case when the original
  lemma instance was inside an [encapsulate] event. Now, that
  information is stored provided the encapsulate event has an empty
  [signature] list and is not ``empty'' (that is, stores at least one
  event). Thanks to Eric Smith for sending us an example to point out
  this problem.

Moved the definition of define-sharp-f to avoid a compiler warning
from Allegro CL; not clear if this was truly a bug.

Now, write-list (defined in books/misc/file-io.lisp) and
write-list-downcase take a new keyword argument :quiet, which is nil
by default for the existing behavior, but :quiet t avoids the noise.