Eschew all boiling
Judging by a recent thread on the acl2-help mailing list, it can be
confusing / unhelpful / annoying for users to see the name
BOIL-THAT-FLYSPECK in warning messages printed when theory invariants
from arithmetic-5 are violated.  This commit changes the name to
ARITH-5-ACTIVE-FLAG, since the purpose of the name is to flag (by its
own activeness) whether the arithmetic-5 rules are active.