Display cert.pl build times to the millisecond
A lot of books certify in "0s" or "1s", so I thought it would be nice to
offer a little higher precision for ranking purposes.  Millisecond
precision is also the default output for the `time` command in recent
Bash versions.

The Time::HiRes module is shipped with the base Perl system since Perl
5.8, which was released in 2002, so I think it should be safe to use.