Print non-Boolean masks on a separate line
When printing `svex-env-p` values using `svtv-print-alist-readable`,
every 4vec value which has some non-Boolean component (i.e. is not a
2vec) is (before this commit) displayed with a comment off to the
right giving a "non-Boolean mask", which is a bitvector showing which
bits are either X or Z, i.e. not a Boolean value 0 or 1.

This commit moves the non-Boolean mask to a separate line and aligns
it with the upper and lower components of the 4vec.  With the new
positioning, it's easier to visually compare the non-Boolean mask with
the 4vec's components, as well as with other nearby values of the same
bit width.

Also, this change makes it far less likely for the pretty-printed
svex-env-p value to run into word wrapping issues.  Formerly, due to
the spacing used, the non-Boolean mask would commonly (in my
experience) be wrapped to the next line if the 4vec was 32 bits or
wider and had a non-Boolean MSB.  Now it shouldn't need to wrap until
it's around 160 bits wide, assuming the default right margin of around
77 columns.

Before:

```lisp
  (FOO                   #ux1_FFFF         ;; non-Boolean mask: #ux1_-
FFFF
                       . #ux0)
```

After:

```lisp
  (FOO                   #ux1_FFFF
                       . #ux0)
;;;    non-Boolean mask: #ux1_FFFF
```