Fixed bugs in tracking and disabling compound-recognizer rules.  Added full :doc for untranslate.  Moved existing items on acl2-doc to the right section.
Quoting :doc note-7-4:

  Fixed bugs in the tracking and disabling of [compound-recognizer]
  rules. Thanks to Dmitry Nadezhin for pointing out this bug, indeed,
  sending an example and analyzing it to point to system function
  term-and-typ-to-lookup, which was indeed the one that had the bug.

Replaced the defpointer for untranslate with a proper xdoc topic.
Thanks to Eric Smith for bringing to our attention this deficiency.

Two items on acl2-doc had been in the section on "Changes at the
System Level", but are now in the more suitable section, "EMACS
Support".