Improved type-prescription rule inference.  Tweaked new xdoc-error handling.  Modified :defthm-name argument of install-not-normalized to be doublets.
Quoting :doc note-7-4:

  A technical change in the handling of [type-set] computations can, in
  rare cases, result in better built-in [type-prescription] rules for
  functions.

Tweaked the new xdoc-error handling.

Modified install-not-normalized so that when the :defthm-name argument
is not a symbol, it is a list of doublets.  So for example, instead of
((f1 . g1) (f2 . g2))
as a value for :defthm-name, instead that value would be
((f1 g1) (f2 g2)).
Thanks to Eric Smith for requesting this change.