Fix bug in GitHub Issue #690.
See https://github.com/acl2/acl2/issues/690

Also slightly extend documentation to clarify that :REWRITE is only allowed when
the quantifier is universal, and to explain that it is inherited by the function
being instantiated when not supplied for the new instance.

Also add the test in the Issue to the regression test suite (a new file that
should be extended with more tests).