Arranged for `make' to be a no-op when no supporting files have changed.  Disabled (:e print-call-history).  Minor :doc tweaks.  Fixed issue in conversion of implicate to :logic mode.
Quoting :doc note-7-4:

  An invocation of `make' to build an ACL2 executable now only
  initiates a build when necessary (that is, when at least one source
  file or supporting file, like GNUmakefile, has changed). Thanks to
  Eric Smith for the suggestion. Removed the following `make'
  targets, which were suspect after this change (the first definitely
  no longer worked) and are unnecessary: saved_acl2, saved_acl2p,
  saved_acl2r, and saved_acl2pr. Note that now, a successful build
  shows only four lines for 'fgrep -i compiler make.log' instead of
  five.

Note: Part of the change above is that TAGS-acl2-doc will not be
rebuilt except when sources (or supporting files like GNUmakefile)
change.

Quoting a comment in (defxdoc note-7-4 ...):

; Disabled executable-counterpart of function print-call-history; so now we
; don't get distracting output, for example from
; (thm (equal (print-call-history) 3)).

Added to :doc note-7-4 in a way that will be in future such :doc
topics as well (motivated by recent conversion of implicate to :logic
mode):

  .... Also note that
  with each release, some built-in functions that were formerly in
  :[program] mode are now see guard-verified :[logic] mode functions.

Fixed typo in :doc user-defined-functions-table.

Fixed "make devel-check" issue for implicate, by arranging, in
books/system/kestrel.lisp, to verify termination and guards for
dumb-negate-lit before doing so for implicate.