Improve TYPED-TUPLEP macro.
In its previous form, this macro was useful for execution but not for rewrite
rules, due to the use of = instead of EQUAL and to the use of a final T conjunct
in order to have the macro always return a boolean.

The use of = has been replaced with EQUAL, and the final T conjunct has been
removed (allowing the macro to return a non-boolean).