Make `find-lemmas` much faster
`find-lemmas-fn` was repeatedly calling `deref-macro-name` on the
function names supplied by the user, every time it inspected an event
recorded in the world.  When the world is very long, this can be very
time consuming because `deref-macro-name` uses tables, which as
pointed out in `:doc tables` are "not especially efficient".

This commit makes it so that `deref-macro-name-list` is only called
once per invocation of `find-lemmas`.  In one example, running
`find-lemmas` in a world containing more than 1.5 million triples used
to take more than 5 minutes, but now it completes in 0.15 seconds.