Modified certification procedure for books/projects/apply-model/.
Thanks to Eric Smith for pointing out warnings that led me to re-work
the infrastructure for certifying the apply-model/ books, and thanks
to Sol Swords for adding the new --include-excludes option for cert.pl
in support of this change.

Fixed a few other related things while I was at it.