Extend utilities for event forms.
A new utility produces one of the symbols DEFUN, DEFUND, DEFUN-NX, and DEFUND-NX
from two input flags for enablement and non-executability.

Another new utility produces one of the symbols DEFTHM and DEFTHMD from an input
flag for enablement.

These are useful, for instance, in code that generates function and theorem
event forms.