VL: fix parsing bug with property formals
Formal parameters to properties can have a datatype, or else they
can be plain, untyped identifiers.  We weren't handling the
second case correctly because we were always expecting to parse a
data_type_or_implicit, which can suffer from the usual problem of
a plain identifier looking like a named data type.  Fix this in
the usual way: look for the datatype and subsequent identifier,
and if that fails then match a plain id.