A utility for untranslation preprocessing.
This utility provides a macro to keep a constant unexpanded in the ACL2 output.

The macro updates the UNTRANSLATE-PREPROCESS entry of the
USER-DEFINED-FUNCTIONS-TABLE with a new function to replace the constant's
expansion with its name and to delegate all other terms to the previously
installed function (if any). This macro can be used after introducing a constant
that the user wants to keep unexpanded in the output.