Add ABNF (Augmented Backus-Naur Form) library.
ABNF is a a standardized formal grammar notation used in several Internet syntax
specifications, e.g. URI, HTTP, IMF, SMTP, IMAP, and JSON.

ABNF adds conveniences (e.g. repetition prefixes, similar to regular
expressions) and makes slight modifications to BNF (Backus-Naur Form), without
going beyond context-free grammars.

This ACL2 library provides:
1. A formalization of the syntax and semantics of the ABNF notation.
2. A verified parser that turns ABNF grammar text (e.g. copied and pasted from
   the standard that defines HTTP) into a formal representation suitable for
   formal specification (e.g. for correct HTTP parsing).
3. Executable operations on ABNF grammars, e.g. to check their validity and to
   compose them.
The library is thoroughly documented using XDOC, particularly the parser
correctness proof.

The completeness part of the proof (i.e. the parser recognizes all valid
grammars) is considerably more laborious than the soundness part of the proof
(i.e. the parser recognizes only valid grammars). The completeness proof
involves showing that the parser can successfully disambiguate the various
alternatives in the grammar rules of ABNF. It turns out that the grammar rules
of ABNF (from the ABNF standard) are slightly ambiguous in their treatment of
comments and white space in the top-level rule; this is proved via a witnessing
example. The parser resolves this ambiguity (which does not affect the ABNF
semantics, anyhow) in a precisely characterized way, and completeness is
(necessarily) proved relatively to that disambiguating restriction. Aside from
this ambiguity, the rest of the grammar of ABNF is mostly LL(1), with a few
LL(2) and LL(*) portions.  The parser is implemented as recursive descent with
backtracking, using the Seq macros from the Community Books; backtracking is
expected to be limited in reasonable grammars, and in fact the parser runs
quickly on the grammars of HTTP and of the other Internet standards mentioned
above.

Since the ABNF syntax (which the parser parses) is itself specified in ABNF in
the ABNF standard, the verification techniques used for this parser should be
applicable to other parsers of ABNF-specified syntax (e.g. an HTTP parser).

The meta-circularity of the ABNF syntax being defined in ABNF, is broken by
first formalizing an abstract syntax of ABNF in ACL2 (using the FTY library from
the Community Books), and then using this abstract syntax to transcribe the
concrete syntax of ABNF from the ABNF standard to ACL2. After that, it is also
shown that running the parser on the grammar text from the ABNF standard yields
the same result as the manually transcribed one, providing some validation.