Make cert.pl try harder to find the system books
Currently, cert.pl looks for the system books in the following
locations, in order, until one succeeds (defaulting to the second
one):

1) the `books` subdirectory of the directory containing the ACL2
   executable cert.pl is using

2) the parent directory of the directory where cert.pl itself lies

This commit adds another option in between the two:

1.5) the value of `(@ system-books-dir)` reported by the ACL2
     executable when it is run

This is useful if, for example, your ACL2 executable is actually a
script that sets environment variables or does other kinds of setup
before actually calling ACL2 itself, and that script resides in a
totally different location from the actual ACL2 image that it
eventually launches.

This commit also cleans up a couple of comments, and adds an output
warning message in the case that the system books directory could not
be determined, where previously cert.pl would silently continue.