VL: improve memory usage for with-ps-file
Previously, with-ps-file built the full string to print and then wrote it out
to file all at once, which can use a lot of memory.  Change this to use an
incremental writer.  Use some under-the-hood stuff to make it fast (nreverse
and bypass princ$ since it is quite slow versus write-char/write-string).