Add utilities for programmatic error checking.
These utilities serve to concisely check for erroneous conditions (e.g. on
inputs from a user to a macro) and generate a soft error, with an informative
error message, if the conditions are not satisfied. Using these utilities help
keep the error checking code concise and readable, and keep the user interface
(error messages) uniform.

These utilities also include a macro DEF-ERROR-CHECKER to define error-checking
functions of a certain form. Most of the functions in these utilities are
generated via this macro.